+++ site-cookbooks/alterf/recipes/rc.local.rb
@@ -6,9 +6,22 @@
 #
 # All rights reserved - Do Not Redistribute
 #
-template '/etc/rc.d/rc.local' do
+file '/etc/rc.d/rc.local' do
+  content IO.read(File.absolute_path(File.dirname(__FILE__) + '/../files/default/rc.local'))
+  owner 'root'
+  group 'root'
+  mode 0755
+end
+
+directory '/etc/rc.d/rc.local.d' do
+  user 'root'
+  group 'root'
+  mode 0755
+end
+
+template "/etc/rc.d/rc.local.d/#{node[:alterf][:app_name]}" do
   source 'rc.local.erb'
   owner 'root'
   group 'root'
-  mode 0775
+  mode 0755
 end